+++ examples/apps/web-components-ssx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web-components-ssx
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+  - @inglorious/ssx@2.1.2
+
 ## 0.1.15
 
 ### Patch Changes