+++ examples/apps/web-components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-components
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+
 ## 0.0.11
 
 ### Patch Changes