+++ examples/apps/web-finance-dashboard/CHANGELOG.md
@@ -1,5 +1,14 @@
 # web-finance-dashboard
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+  - @inglorious/charts@8.0.0
+  - @inglorious/ui@5.0.0
+
 ## 0.1.4
 
 ### Patch Changes