+++ examples/apps/web-todomvc-ace/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-todomvc-ace
 
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+
 ## 0.1.23
 
 ### Patch Changes