+++ examples/apps/web-todomvc-jsx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-todomvc-jsx
 
+## 0.1.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+
 ## 0.1.33
 
 ### Patch Changes