+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/charts
 
+## 8.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+
 ## 7.0.3
 
 ### Patch Changes