+++ packages/motion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/motion
 
+## 7.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+
 ## 6.0.2
 
 ### Patch Changes