+++ packages/ssx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/ssx
 
+## 2.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+
 ## 2.1.1
 
 ### Patch Changes