+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inglorious/ui
 
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.2.0
+  - @inglorious/charts@8.0.0
+
 ## 4.0.3
 
 ### Patch Changes